# api-backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IntelliQ API",
        "version": "0.0.1"
    },
    "paths": {
...
        "/intelliq_api/question/{questionnaireID}/{questionID}": {
            "get": {
                "summary": "Get data for a specific question",
                "description": "Returns data for a specific question.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "questionnaireID",
                        "required": true,
                        "description": "ID of the questionnaire the question belongs to"
                    },
                    {
                        "in": "path",
                        "name": "questionID",
                        "required": true,
                        "description": "ID of the question"
                    },
                    {
                        "in": "query",
                        "name": "format",
                        "description": "Format for the returned data (either \"csv\" or \"json\")"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully returned data for the specified question",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "questionnaireID": {
                                            "type": "string",
                                            "description": "ID of the questionnaire the question belongs to"
                                        },
                                        "qID": {
                                            "type": "string",
                                            "description": "ID of the question"
                                        },
                                        "qtext": {
                                            "type": "string",
                                            "description": "Text of the question"
                                        },
                                        "required": {
                                            "type": "string",
                                            "description": "Indicates whether the question is required (either \"true\" or \"false\")"
                                        },
                                        "type": {
                                            "type": "string",
                                            "description": "Type of the question (either \"question\" or \"profile\")"
                                        },
                                        "options": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "optID": {
                                                        "type": "string",
                                                        "description": "ID of the option"
                                                    },
                                                    "opttxt": {
                                                        "type": "string",
                                                        "description": "Text of the option"
                                                    },
                                                    "nextqID": {
                                                        "type": "string",
                                                        "description": "ID of the next question"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No questionnaireID or questionID was provided"
                    },
                    "402": {
                        "description": "No data was found for the specified questionnaireID and questionID"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/intelliq_api/questionnaire/{questionnaireID}": {
            "get": {
                "summary": "Get questionnaire information",
                "description": "Returns the information of a specific questionnaire",
                "parameters": [
                    {
                        "name": "questionnaireID",
                        "in": "path",
                        "required": true,
                        "description": "ID of the questionnaire",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "questionnaireID": {
                                            "type": "string"
                                        },
                                        "questionnaireTitle": {
                                            "type": "string"
                                        },
                                        "keywords": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "questions": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "qid": {
                                                        "type": "string"
                                                    },
                                                    "qtext": {
                                                        "type": "string"
                                                    },
                                                    "required": {
                                                        "type": "string",
                                                        "enum": [
                                                            "true",
                                                            "false"
                                                        ]
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "profile",
                                                            "question"
                                                        ]
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing required parameter: questionnaireID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Missing required parameter: questionnaireID"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "No data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No data"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Internal error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "produces": [
                    "application/json",
                    "text/csv"
                ]
            }
        },
...
        "/intelliq_api/admin/questionnaire_upd": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Update a questionnaire",
                "description": "Updates the questionnaire details in the database",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "questionnaire",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "questionnaireID",
                                "questionnaireTitle",
                                "keywords",
                                "questions"
                            ],
                            "properties": {
                                "questionnaireID": {
                                    "type": "string",
                                    "description": "Unique identifier for the questionnaire",
                                    "maxLength": 10
                                },
                                "questionnaireTitle": {
                                    "type": "string",
                                    "description": "Title of the questionnaire",
                                    "maxLength": 255
                                },
                                "keywords": {
                                    "type": "array",
                                    "description": "Keywords associated with the questionnaire",
                                    "items": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "questions": {
                                    "type": "array",
                                    "description": "List of questions in the questionnaire",
                                    "items": {
                                        "type": "object",
                                        "required": [
                                            "qID",
                                            "qtext",
                                            "required",
                                            "type",
                                            "options"
                                        ],
                                        "properties": {
                                            "qID": {
                                                "type": "string",
                                                "description": "Unique identifier for the question",
                                                "maxLength": 10
                                            },
                                            "qtext": {
                                                "type": "string",
                                                "description": "Text of the question",
                                                "maxLength": 255
                                            },
                                            "required": {
                                                "type": "string",
                                                "description": "Indicates if the question is required or not",
                                                "enum": [
                                                    true,
                                                    false
                                                ]
                                            },
                                            "type": {
                                                "type": "string",
                                                "description": "Type of the question (profile or question)",
                                                "enum": [
                                                    "profile",
                                                    "question"
                                                ]
                                            },
                                            "options": {
                                                "type": "array",
                                                "description": "List of options for the question",
                                                "items": {
                                                    "type": "object",
                                                    "required": [
                                                        "optID",
                                                        "opttxt",
                                                        "nextqID"
                                                    ],
                                                    "properties": {
                                                        "optID": {
                                                            "type": "string",
                                                            "description": "Unique identifier for the option",
                                                            "maxLength": 10
                                                        },
                                                        "opttxt": {
                                                            "type": "string",
                                                            "description": "Text of the option",
                                                            "maxLength": 255
                                                        },
                                                        "nextqID": {
                                                            "type": "string",
                                                            "description": "Identifier for the next question to be asked based on this option",
                                                            "maxLength": 10
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request (malformed input)"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}